fix: 抽取type
