+++ src/packages/picker/types.ts
@@ -1,0 +1,7 @@
+export interface PickerOption {
+  text: string | number
+  value: string | number
+  disabled?: boolean
+  children?: PickerOption[]
+  className?: string | number
+}
